Refactored the formatting #44
The avatars are now placed just before the title, and the issue ID and
number of comments appear before the avatar.

I'm thinking of create a -make-...-string function for each of the
components that make up the title, and somehow parse that into a
customizable formatting object, e.g.

%ID% %COMMENTS% %AVATAR% %LABELS% %TITLE% etc

And map each of these symbols to a function call with an a-list?

The only difficult part will be placing the avatars as they need to be
done with a separate insert, unless I wrap the magit-insert inside each
of the above functions?