Huge Refactor to improve formatting of titles #54
Users can now customise the format of the issue titles. This implicitly
allows for feature toggling without the need for lots of custom boolean
variables, i.e. if the user doesn't include `%labels%` in
`magit-gh-issues-title-template-string` then there's no need to make the
requests for label information etc, similarly with avatars.

It also removes the need for calculating padding between title elements
as the user can choose this, they can also put their own plain text
items in the title if they want to.